pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autophot-object"
version = "0.1.0"
description = "AutoPHOT object photometry pipeline (CCD/NIR; calibrated aperture/PSF photometry; template subtraction)."
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
authors = [{name = "AutoPHOT contributors"}]
dependencies = [
  "aafitrans==0.2.1",
  "astroalign==2.6.1",
  "astropy==6.1.6",
  "astroquery==0.4.7",
  "astroscrappy==1.2.0",
  "emcee==3.1.6",
  "lmfit==1.3.2",
  "mastcasjobs==0.0.7",
  "matplotlib==3.10.0",
  "numpy>=1.26,<2",
  "pandas==2.2.3",
  "photutils==2.0.2",
  "PyYAML==6.0.2",
  "reproject==0.14.1",
  "requests==2.32.3",
  "scikit-image==0.25.0",
  "scikit-learn>=1.3",
  "scipy==1.15.1",
  "sfft==1.6.3",
]

[project.scripts]
autophot-main = "main:run_photometry"
autophot-driver = "autophot:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["utils*", "databases*"]

[tool.setuptools.py-modules]
modules = [
  "aperture",
  "autophot",
  "autophot_tokens",
  "background",
  "batch_main",
  "catalog",
  "check",
  "cosmic",
  "fix_template_headers",
  "functions",
  "fwhm",
  "hst",
  "lightcurve",
  "limits",
  "main",
  "mask",
  "move_archive3_templates",
  "move_templates",
  "plot",
  "prepare",
  "psf",
  "templates",
  "tns",
  "wcs",
  "zeropoint",
]

[tool.setuptools.package-data]
databases = ["*.yml"]
utils = ["*.py"]

[tool.setuptools.data-files]
"." = ["autophot.mplstyle", "gauss_3.0_7x7.conv", "gaussian_7x7.conv"]

